Explicitly add project.scm.tag
The release plugin (specifically the various AbstractRewritePomsPhase-s) will
set the tag for the generated POMs, which introduces "weird" formatting. This
might be ok for the release POM, which isn't intended for human consumption,
but unfortunately the plugin (no longer?) removes the value when building
the "next development iteration" POM.

By having the element here the plugin will actually modify only the value, leaving
the indentation alone.

The maven release process using the maven-release-plugin is essentially brain-dead,
so when using this POM as a parent the complete project.scm block should be copied
as-is into the child POM.